reenable angular-mocks tests
A little setTimeout always helps, right?

I think the problem was in which sequence things happen. Before the transform
refactoring in in traverson (commit 418abd5365a5ab4dc605670e54e2e1d4a3fe9878)
everything might have happened synchronously in this test, because with
angular-mocks, the callback for the http calls are called immediately,
synchronously. Now I *guess* there is a forced asynchronicity in the
call flow and so the httpBackend.flush() call came to early so that the
second http request came actually after the flush and was not responded
to.

Closes #12